feat: add initial model, repository, service, and controller for Echange, CategorieObjet, and Utilisateur
