+++ Backend/src/main/java/fr/knap/testunitaire_esiee/model/CategorieObjet.java
