+++ Backend/src/main/java/fr/knap/testunitaire_esiee/respository/Interaction.java
@@ -1,0 +1,25 @@
+package fr.knap.testunitaire_esiee.respository;
+
+import fr.knap.testunitaire_esiee.model.Objet;
+
+import java.util.List;
+
+public abstract class Interaction {
+
+    /**
+     * Default constructor
+     */
+    public Interaction() {
+    }
+
+    /**
+     * @param object
+     * @return
+     */
+    public abstract Object ConsulterObjet(Object object);
+
+    /**
+     * @return
+     */
+    public abstract List<Objet> RechercheObjet();
+}
