+++ Backend/src/main/java/fr/knap/testunitaire_esiee/services/UtilisateurService.java
@@ -1,0 +1,37 @@
+package fr.knap.testunitaire_esiee.services;
+
+import fr.knap.testunitaire_esiee.model.Utilisateur;
+import fr.knap.testunitaire_esiee.respository.UtilisateurRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class UtilisateurService {
+    @Autowired
+    private UtilisateurRepository utilisateurRepository;
+
+    public Utilisateur creerUtilisateur(Utilisateur utilisateur) {
+        return utilisateurRepository.save(utilisateur);
+    }
+
+    public List<Utilisateur> obtenirTousLesUtilisateurs() {
+        return utilisateurRepository.findAll();
+    }
+
+    public Utilisateur obtenirUtilisateurParId(Long id) {
+        return utilisateurRepository.findById(id).orElse(null);
+    }
+
+    public Utilisateur mettreAJourUtilisateur(Long id, Utilisateur utilisateur) {
+        if (utilisateurRepository.existsById(id)) {
+            return utilisateurRepository.save(utilisateur);
+        }
+        return null;
+    }
+
+    public void supprimerUtilisateur(Long id) {
+        utilisateurRepository.deleteById(id);
+    }
+}
