+++ Backend/src/test/java/fr/knap/testunitaire_esiee/model/UtilisateurTest.java
