+++ docker-compose.yml
@@ -1,9 +1,8 @@
 services:
-  app:
+  backend:
     build:
-        context: .
-        dockerfile: Dockerfile
-    container_name: Test-Unitaire_Esiee
+      context: ./Backend
+      dockerfile: Dockerfile
     ports:
       - "3000:3000"
     environment:
@@ -11,3 +10,11 @@ services:
       - SPRING_DATASOURCE_USERNAME=sa
       - SPRING_DATASOURCE_PASSWORD=password
       - SPRING_JPA_HIBERNATE_DDL_AUTO=update
+
+  frontend:
+    build:
+      context: ./Front
+      dockerfile: Dockerfile
+    ports:
+      - "8080:8080"
+      - "80:8080"
